URI p-1131 wrong answer 5%

+++ URI_midterm/p-1021/src/Main.java
@@ -10,25 +10,25 @@ public class Main {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
 
-        float n = s.nextFloat();
+        double n = s.nextDouble();
 
         System.out.print("NOTAS:\n");
         System.out.printf("%d nota(s) de R$ 100.00\n", (int)(n/100.00));
-        n%=100.00;
+        n%=100;
         System.out.printf("%d nota(s) de R$ 50.00\n", (int)(n/50.00));
-        n%=50.00;
+        n%=50;
         System.out.printf("%d nota(s) de R$ 20.00\n", (int)(n/20.00));
-        n%=20.00;
+        n%=20;
         System.out.printf("%d nota(s) de R$ 10.00\n", (int)(n/10.00));
-        n%=10.00;
+        n%=10;
         System.out.printf("%d nota(s) de R$ 5.00\n", (int)(n/5.00));
-        n%=5.00;
+        n%=5;
         System.out.printf("%d nota(s) de R$ 2.00\n", (int)(n/2.00));
-        n%=2.00;
+        n%=2;
 
         System.out.print("MOEDAS:\n");
         System.out.printf("%d moeda(s) de R$ 1.00\n", (int)(n/1.00));
-        n%=1.00;
+        n%=1;
         System.out.printf("%d moeda(s) de R$ 0.50\n", (int)(n/0.5));
         n%=0.50;
         System.out.printf("%d moeda(s) de R$ 0.25\n", (int)(n/0.25));
